refactor: move ethers related file into /src/modules/ethers
