+++ src/components/display/mint/MintSection.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import { Card, Row, Col, Spinner } from "react-bootstrap";
 import MintCard from "./MintCard";
-import { getMintPositions, MintPosition } from "../../../modules/Loan";
-import { getAddressInQueryString } from "../../../modules/Utils";
+import { getMintPositions, MintPosition } from "../../../modules/ethers/Loan";
+import { getAddressInQueryString } from "../../../modules/ethers/Utils";
 import IconTooltip from "../../common/IconTooltip";
 
 const MintSection = () => {
